feat: New experiment tracking and management system
This commit introduces a new experiment tracking and management system
that allows users to create, delete, and select experiments.  The
system also integrates with the notebook interface, allowing users to
easily switch between experiments and view results.

The experiment interface now uses a dropdown for experiment
selection and a new experiment directory is used for generated results.

The dropdown allows users to select from a list of existing
experiments or create a new one. The new experiment directory is used
to store the generated results for the selected experiment.

Also converted to use the pathlib Path class for accessing output
directories, up until the calls to TimeloopFE, which does not
understand Posit paths.